+++ ElvUI_MerathilisUI/Core/Changelog/6.11.lua
@@ -1,6 +1,8 @@
 local MER, F, E, I, V, P, G, L = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[611] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "13.04.2004",
+	FIXES = {
+		"[Core]: Fix an issue with Windtools enabled",
+	},
 }