+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 ### Changes ###
 
-  * [Core]: Fix a a whoospi with changelog option path
-  * [RaidManager]: Fix errors in the RaidManager that prevent him from loading
+  * [Core]: Fix an issue with WindTools enabled